+++ pysat/tests/classes/cls_instrument_access.py
@@ -786,6 +786,7 @@ class InstAccessTests(object):
         self.testInst.load(self.ref_time.year, self.ref_doy, use_header=True)
         self.testInst['doubleMLT'] = 2.
         assert np.all(self.testInst['doubleMLT'] == 2.)
+        assert len(self.testInst['doubleMLT']) == len(self.testInst.index)
 
         self.testInst['nanMLT'] = np.nan
         assert np.all(np.isnan(self.testInst['nanMLT']))